IOS-3109 Refactor the app's start
